Add topic reply

+++ src/Controller/TopicController.php
@@ -3,7 +3,9 @@
 namespace App\Controller;
 
 use App\Entity\Category;
+use App\Entity\Post;
 use App\Entity\Topic;
+use App\Form\NewPostType;
 use App\Form\NewTopicType;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -31,6 +33,33 @@ class TopicController extends AbstractController
         return $this->render('topic/index.html.twig', [
             'topic' => $topic,
             'posts' => $posts
+        ]);
+    }
+
+    #[Route('/topic/{id}/reply', name: 'reply_topic')]
+    public function reply(Topic $topic, Request $request, ManagerRegistry $managerRegistry, PaginatorInterface $paginatorInterface): Response
+    {
+        $post = new Post();
+        $post->setTopic($topic);
+        $post->setAuthor($this->getUser());
+
+        $form = $this->createForm(NewPostType::class, $post);
+        $form->handleRequest($request);
+        
+        if ($form->isSubmitted() && $form->isValid()) {
+            
+            $entityManager = $managerRegistry->getManager();
+            $entityManager->persist($post);
+            $entityManager->flush();
+            
+            return $this->redirectToRoute('topic', [
+                'id' => $topic->getId()
+            ]);
+        }
+
+        return $this->render('post/create.html.twig', [
+            'topic' => $topic,
+            'form' => $form->createView()
         ]);
     }
     
@@ -66,7 +95,7 @@ class TopicController extends AbstractController
     {
         $categoryId = $topic->getCategory()->getId();
         
-        if($this->getUser() == $topic->getAuthor()) {
+        if($this->getUser() == $topic->getAuthor() && $topic->getPosts()->count() <= 0) {
             $entityManager = $managerRegistry->getManager();
             $entityManager->remove($topic);
             $entityManager->flush();
